Fixed a potential bug.
If one or more threads where trying to disconnect a slot that was being
destructed in another thread, the result could be that the disconnecting
connection was using a destructed disconnector instance. This is now avoided
by stalling the constructor until all weak pointers to the disconnector have
no active shared pointer aquired.