Extracted the thread syncronization to a policy.
The signals previously implemented thread synchronization using `std::mutex`
and `std::lock_guard`. This has been moved into a thread policy that is now
a template parameter to the signal template.

The signal template has therefore changed name to `signal_type`, and the
template `signal` is now a template alias for using the `signal_type` template
and the multithread policy.

Added a new type of signal that is not safe for multithreaded use. This uses
a single thread policy that implements the mutex and lock types as dummys. This
new type is called `nod::unsafe_signal`