Slot return values.
Added the possibility to create a signal with a collector type. The collector
will be instantiated when triggering a signal, and it will be passed the
return values of all called slots. The collector result will be returned
from the signals function call operator.

The collector must provide a type definition of a `result_type` and must
implement the function operator.

A simple collector can look like this:

```cpp
template <class Container>
struct aggregate
{
    using result_type = typename Container;
    void operator()( Container& partial_result,
                     typename Container::value_type&& next_result ) const {
        partial_result.push_back( std::move(next_result) );
    }
};
```

This collector will aggregate all slot return values by pushing them into
a container.